Move timeline data generation to a dedicated file
