+++ rhods-ci/plotting/advanced_timeline.py
@@ -8,6 +8,8 @@ import plotly.express as px
 import matrix_benchmarking.plotting.table_stats as table_stats
 import matrix_benchmarking.common as common
 
+from . import timeline_data
+
 def register():
     Timeline("Timeline")
 
@@ -22,243 +24,22 @@ class Timeline():
     def do_hover(self, meta_value, variables, figure, data, click_info):
         return "nothing"
 
-    def do_plot(self, ordered_vars, params, param_lists, variables, cfg):
+    def do_plot(self, ordered_vars, settings, setting_lists, variables, cfg):
 
-        all_XY = defaultdict(dict)
-        user_count = 0
-
-        if sum(1 for _ in common.Matrix.all_records(params, param_lists)) != 1:
+        if sum(1 for _ in common.Matrix.all_records(settings, setting_lists)) != 1:
             return {}, "ERROR: only one experiment must be selected"
 
-        def get_text(what, start_evt, finish_evt, start, finish):
-            duration_s = (finish - start).total_seconds()
-
-            if duration_s < 120:
-                duration = f"{duration_s:.1f} seconds"
-            else:
-                duration_min = duration_s / 60
-                duration = f"{duration_min:.1f} minutes"
-
-            if finish_evt:
-                return f"{what}<br>FROM: {start_evt} ({start})<br>TO: {finish_evt} ({finish})<br>{duration}"
-            else:
-                return f"{what}<br>{start_evt}<br>{duration}"
-
+        user_count = 0
         data = []
-        for entry in common.Matrix.all_records(params, param_lists):
-            user_count = len(entry.results.test_pods)
-            test_nodes = {}
-            rhods_nodes = {}
-
-            test_nodes_index = list(entry.results.testpod_hostnames.values()).index
-            rhods_nodes_index = list(entry.results.notebook_hostnames.values()).index
-
-            for testpod_name, nodename in entry.results.testpod_hostnames.items():
-                user_idx = int(testpod_name.split("-")[2])
-                test_nodes[user_idx] = f"Test node #{test_nodes_index(nodename)}"
-
-            for notebook_name, nodename in entry.results.notebook_hostnames.items():
-                user_idx = int(notebook_name.replace("jupyterhub-nb-testuser", ""))
-                rhods_nodes[user_idx] = f"RHODS node #{rhods_nodes_index(nodename)}"
-
-            def get_line_name(user_idx):
-                line_name = [
-                    f"User #{user_idx}",
-                    test_nodes[user_idx],
-                ]
-
-                if rhods_nodes:
-                    line_name += [rhods_nodes.get(user_idx, "<no RHODS node>")]
-
-                return user_name + "<br>" + test_node + "<br>" + rhods_node
-
-            for testpod_name in entry.results.test_pods:
-                user_idx = int(testpod_name.split("-")[2])
-
-                pod_times = entry.results.pod_times[testpod_name]
-
-                event_times = entry.results.event_times[testpod_name]
-
-                if not pod_times.__dict__: continue
-
-                def generate_data(LegendName, start_evt, finish_evt, **kwargs):
-                    defaults = dict(
-                        LegendName=LegendName,
-                        LegendGroup="Test Pod",
-                        Start=kwargs.get("Start") or data[-1]["Finish"],
-                        Finish=kwargs["Finish"],
-                        LineName=get_line_name(user_idx),
-                        Opacity=0.5,
-                        LineWidth=80,
-                        LineSortIndex=kwargs["Finish"],
-                    )
-
-                    defaults["Text"] = get_text(LegendName,
-                                                start_evt, finish_evt,
-                                                defaults["Start"], defaults["Finish"])
-
-                    return defaults | kwargs
-
-                data.append(generate_data(
-                    "01. Test pod scheduling",
-                    "Job creation", "Pod scheduled",
-                    Finish=event_times.scheduled,
-                    Start=entry.results.job_creation_time,))
-                data.append(generate_data(
-                    "02. Test pod preparation",
-                    "Pod scheduled", "pulling image",
-                    Finish=event_times.pulling))
-                data.append(generate_data(
-                    "03. Test pod image pull",
-                    "Pod pulling image", "image pulled",
-                    Finish=event_times.pulled))
-                data.append(generate_data(
-                    "04. Test pod initialization",
-                    "Pod image pulled", "container started",
-                    Finish=pod_times.container_started))
-                data.append(generate_data(
-                    "05. Test Execution",
-                    "Container started", "container finished",
-                    Finish=pod_times.container_finished))
-
-                for reason, start, end, msg in event_times.__dict__.get("warnings", []):
-                    data.append(generate_data(
-                        "Test-pod K8s Warnings",
-                        reason, msg,
-                        Start=start,
-                        Finish=end,
-                        LineColor="Red",
-                        LineWidth=20,
-                        Opacity=0.9,
-                    ))
-
-            data_length_before_ods_ci = len(data)
-
-            for testpod_name, ods_ci_output in entry.results.ods_ci_output.items():
-                user_idx = int(testpod_name.split("-")[2])
-
-                for test_idx, (test_name, test_times) in enumerate(ods_ci_output.items()):
-                    def generate_data(LegendName, status, **kwargs):
-                        defaults = dict(
-                            LegendName=LegendName,
-                            LegendGroup="ODS-CI",
-                            Start=kwargs.get("Start") or data[-1]["Finish"],
-                            Finish=kwargs["Finish"],
-                            LineName=get_line_name(user_idx),
-                            Opacity=0.9,
-                            LineWidth=50,
-                            LineSortIndex=kwargs["Finish"],
-                        )
-
-                        defaults["Text"] = get_text(LegendName,
-                                                    f"Test result: {status}<br>FROM: {defaults['Start']}<br>TO: {defaults['Finish']}", None,
-                                                    defaults["Start"], defaults["Finish"])
-
-                        return defaults | kwargs
-
-                    data.append(generate_data(
-                        f"ODS - {test_idx} - {test_name}",
-                        test_times.status,
-                        Start=test_times.start,
-                        Finish=test_times.finish,
-                    ))
-
-            for notebook_name in entry.results.notebook_pods:
-                user_idx = int(notebook_name.replace("jupyterhub-nb-testuser", ""))
-
-                pod_times = entry.results.pod_times[notebook_name]
-                event_times = entry.results.event_times[notebook_name]
-
-                appears_time = min([v for v in event_times.__dict__.values() if isinstance(v, datetime.datetime)])
-
-                def generate_data(LegendName, start_evt, finish_evt, **kwargs):
-                    defaults = dict(
-                        Start=kwargs.get("Start") or data[-1]["Finish"],
-                        Finish=kwargs["Finish"],
-                        LegendName=LegendName,
-                        LegendGroup="Notebook",
-                        Opacity=1,
-                        LineName=get_line_name(user_idx),
-                        LineWidth=30,
-                        LineSortIndex=kwargs["Finish"],
-                    )
-
-                    defaults["Text"] = get_text(LegendName,
-                                                start_evt, finish_evt,
-                                                defaults["Start"], defaults["Finish"])
-
-                    return defaults | kwargs
-
-                data.append(generate_data(
-                    "20. Notebook scheduling",
-                    "Notebook pod appeared", "scheduled",
-                    Start=appears_time,
-                    Finish=event_times.scheduled))
-                data.append(generate_data(
-                    "21. Notebook preparation",
-                    "Notebook pod cheduled", "image pulling",
-                    Finish=event_times.pulling))
-                data.append(generate_data(
-                    "22. Notebook image pull",
-                    "Notebook image pulling", "image pulled",
-                    Finish=event_times.pulled))
-                data.append(generate_data(
-                    "23. Notebook initialization",
-                    "Notebook image pulled", "container started",
-                    Finish=event_times.started))
-                data.append(generate_data(
-                    "24. Notebook execution",
-                    "Notebook container started", "container terminated",
-                    Finish=event_times.terminated))
-
-
-
-                for reason, start, end, msg in event_times.__dict__.get("warnings", []):
-                    data.append(generate_data(
-                        "Notebook K8s Warnings",
-                        reason, msg,
-                        Start=start,
-                        Finish=end,
-                        LineColor="Red",
-                        LineWidth=20,
-                        Opacity=0.9,
-                    ))
-
-            for notebook_name in entry.results.notebook_hostnames.keys():
-                user_idx = int(notebook_name.replace("jupyterhub-nb-testuser", ""))
-                rhods_node = rhods_nodes[user_idx]
-
-                LineName = get_line_name(user_idx)
-
-                data.insert(0, dict(
-                    LegendName=rhods_node,
-                    Start=entry.results.job_creation_time - datetime.timedelta(minutes=1),
-                    Finish=entry.results.job_completion_time + datetime.timedelta(minutes=1),
-                    LineName=LineName,
-                    LineWidth=25,
-                    Opacity=0.2,
-                    Text=rhods_node,
-                    LegendGroup="Nodes",
-                    SkipFromMinMaxDate=True,
-                ))
-
-        data.insert(0, dict(
-            LegendName="Test job lifespan",
-            Start=entry.results.job_creation_time,
-            Finish=entry.results.job_completion_time,
-            Opacity=1,
-            Text="Test job lifespan",
-            LineName="Test Job",
-            LineWidth=100,
-        ))
+        line_sort_name = ""
+        for entry in common.Matrix.all_records(settings, setting_lists):
+            user_count, data, line_sort_name = timeline_data.generate(entry)
 
         # ---
 
-        plots = []
         df = pd.DataFrame(data)
-        LINE_SORT_LEGENT_NAME = "05. Test Execution"
 
-        ordered_lines = list(df[df["LegendName"] == LINE_SORT_LEGENT_NAME].sort_values("LineSortIndex")["LineName"].unique())
+        ordered_lines = list(df[df["LegendName"] == line_sort_name].sort_values("LineSortIndex")["LineName"].unique())
         for line_name in df["LineName"].unique():
             if line_name in ordered_lines:
                 continue
@@ -266,6 +47,7 @@ class Timeline():
 
         min_date = []
         max_date = []
+        plots = []
         for legend_name in df["LegendName"].unique():
             rows = df[df["LegendName"] == legend_name]
